requirements.txt
astroid==2.3.3
cycler==0.10.0
decorator==4.4.2
isort==4.3.21
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
matplotlib==3.2.1
mccabe==0.6.1
networkx==2.4
numpy==1.18.2
pylint==2.4.4
pyparsing==2.4.6
python-dateutil==2.8.1
python-igraph==0.8.0
six==1.14.0
texttable==1.6.2
typed-ast==1.4.1
wrapt==1.11.2
